[refactor] Schedule package 경로 변경
상위 경로 ppm -> domain
