+++ src/main/java/org/omoknoone/ppm/domain/schedule/controller/ScheduleController.java
@@ -1,16 +1,16 @@
-package org.omoknoone.ppm.schedule.controller;
+package org.omoknoone.ppm.domain.schedule.controller;
 
 import java.util.List;
 
 import org.modelmapper.ModelMapper;
 import org.modelmapper.TypeToken;
 import org.modelmapper.convention.MatchingStrategies;
-import org.omoknoone.ppm.schedule.aggregate.Schedule;
-import org.omoknoone.ppm.schedule.dto.NewScheduleDTO;
-import org.omoknoone.ppm.schedule.dto.ScheduleDTO;
-import org.omoknoone.ppm.schedule.service.ScheduleService;
-import org.omoknoone.ppm.schedule.vo.RequestSchedule;
-import org.omoknoone.ppm.schedule.vo.ResponseSchedule;
+import org.omoknoone.ppm.domain.schedule.aggregate.Schedule;
+import org.omoknoone.ppm.domain.schedule.dto.NewScheduleDTO;
+import org.omoknoone.ppm.domain.schedule.dto.ScheduleDTO;
+import org.omoknoone.ppm.domain.schedule.service.ScheduleService;
+import org.omoknoone.ppm.domain.schedule.vo.RequestSchedule;
+import org.omoknoone.ppm.domain.schedule.vo.ResponseSchedule;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
