+++ src/main/java/org/omoknoone/ppm/domain/schedule/repository/ScheduleRepository.java
@@ -1,13 +1,10 @@
-package org.omoknoone.ppm.schedule.repository;
+package org.omoknoone.ppm.domain.schedule.repository;
 
-import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Optional;
 
-import org.omoknoone.ppm.schedule.aggregate.Schedule;
+import org.omoknoone.ppm.domain.schedule.aggregate.Schedule;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 
 public interface ScheduleRepository extends JpaRepository<Schedule, Long> {
