+++ src/main/java/org/omoknoone/ppm/domain/schedule/vo/ResponseSchedule.java
@@ -1,4 +1,4 @@
-package org.omoknoone.ppm.schedule.vo;
+package org.omoknoone.ppm.domain.schedule.vo;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
